fix metadata
