+++ src/app/blog/lsm_trees_in_postgres/content.tsx
@@ -1,0 +1,7 @@
+"use client";
+
+import BlogContent from "../../../components/markdown/lsm_trees_in_postgres.mdx";
+
+export default function Content() {
+  return <BlogContent />;
+}
