+++ src/components/markdown/lsm_trees_in_postgres.mdx
@@ -18,9 +18,9 @@ import HotStandbyFeedbackImage from "../../../public/blog/hot_standby_feedback.s
   <div className="relative top-8 h-7 w-7 rounded-full bg-indigo-600 border-2 border-indigo-300 text-white text-xs text-center p-0 right-4 font-semibold">
     <div className="relative top-1">+2</div>
   </div>
-  <p className="relative bottom-2">
+  <div className="relative top-3">
     By Stu Hood, Ming Ying, Mathew Pregasen, and Olive Ratliff on June 30, 2025
-  </p>
+  </div>
 </div>
 
 <Image src={LsmTreeBannerImage} className="rounded-xl" />
